dev/backend/app/tools/mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.3"
dependencies = [
    "fastmcp>=2.8.0",
    "qdrant-client>=1.14.2",
]
